requirements.txt
beautifulsoup4==4.6.3
certifi==2018.8.24
chardet==3.0.4
idna==2.7
mysqlclient==1.3.13
python-dotenv==0.9.1
requests==2.19.1
selenium==3.14.0
SQLAlchemy==1.2.11
urllib3==1.23
